+++ src/types/form.d.ts
@@ -1,0 +1,6 @@
+export interface MultiFormProps {
+    assets: number;
+    buildingType: number;
+    location: string;
+    transactionType: number
+}
